Deleted mock UI for the ContentView
